==> 1_CleanXML.py <==
#!/usr/bin/env python3
# -*- coding: utf-8 -*-
"""
Created on Mon Feb 25 22:38:57 2019

@author: admin
"""

# Step 1 of pre-processing: remove ":dc" from xml elements

import os, fnmatch

def findReplace(directory, find, replace, filePattern):
    for path, dirs, files in os.walk(os.path.abspath(directory)):
        for filename in fnmatch.filter(files, filePattern):
            filepath = os.path.join(path, filename)
            with open(filepath) as f:
                s = f.read()
            s = s.replace(find, replace)
            with open(filepath, "w") as f:
                f.write(s)
    print("Done")
                
findReplace("/../data/xml2", "dc:", "", "*.xml")
